embed java parser source and rename tests
