+++ tests/a2mochi/x/java/for_list_collection.ast
@@ -1,1 +1,10 @@
-
+(program
+  (for n
+    (in
+      (list (int 1) (int 2) (int 3))
+    )
+    (block
+      (call print (selector n))
+    )
+  )
+)
