+++ tests/a2mochi/x/java/string_concat.ast
@@ -1,5 +1,3 @@
 (program
-  (call print
-    (binary + (string "hello ") (string world))
-  )
+  (call print (string "hello world"))
 )
