+++ tools/a2mochi/x/java/print.go
@@ -1,0 +1,42 @@
+//go:build slow
+
+package java
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+	"strings"
+	"time"
+
+	"mochi/ast"
+)
+
+// Print returns Mochi source code for the given AST node with a header.
+func Print(node *ast.Node) (string, error) {
+	if node == nil {
+		return "", fmt.Errorf("nil node")
+	}
+	var b strings.Builder
+	b.WriteString(header())
+	b.WriteString(node.Source())
+	return b.String(), nil
+}
+
+func header() string {
+	tz := time.FixedZone("GMT+7", 7*3600)
+	t := time.Now().In(tz)
+	return fmt.Sprintf("// Generated by a2mochi v%s on %s\n", version(), t.Format("2006-01-02 15:04:05 MST"))
+}
+
+func version() string {
+	root := repoRoot()
+	if root == "" {
+		return "dev"
+	}
+	data, err := os.ReadFile(filepath.Join(root, "VERSION"))
+	if err != nil {
+		return "dev"
+	}
+	return strings.TrimSpace(string(data))
+}
